desquillery_club/service/src/requirements.txt
fastapi
pwntools
uvicorn
typing
gmpy2
pycryptodome
datetime
PyJWT
